sono incappato in un nuovo problema (fondamentamentalmente vado avanti fino a che non giungo ad un problema che mi fa venire voglia di fermarmi per il giorno

+++ music_vs_robot/src/entity/consts.rs
@@ -1,2 +1,2 @@
-pub const MIN_STEP: u32 = 1;
-pub const MAX_STEP: u32 = 5;
+pub const MIN_STEP: usize = 1;
+pub const MAX_STEP: usize = 5;
