+++ music_vs_robot/src/entity/enemy.rs
@@ -3,23 +3,23 @@ use crate::util::visitor::{Visitable, Visitor};
 
 pub trait Enemy: Visitable {
     fn suffer_damage(&mut self, damage: &mut DamagePlayer) -> bool;
-    fn attack(&mut self) -> u32;
-    fn step(&mut self) -> u32;
-    fn value(&self) -> u32;
+    fn attack(&mut self) -> usize;
+    fn step(&mut self) -> usize;
+    fn value(&self) -> usize;
     fn is_dead(&self) -> bool;
     fn copy(&self) -> Box<dyn Enemy>;
 }
 
 #[derive(Clone)]
 pub struct EnemySimple {
-    life: u32,
-    attack: u32,
-    step: u32,
-    value: u32,
+    life: usize,
+    attack: usize,
+    step: usize,
+    value: usize,
 }
 
 impl EnemySimple {
-    pub fn new(life: u32, attack: u32, step: u32, value: u32) -> Self {
+    pub fn new(life: usize, attack: usize, step: usize, value: usize) -> Self {
         EnemySimple {
             life,
             attack,
@@ -47,15 +47,15 @@ impl Enemy for EnemySimple {
         }
     }
 
-    fn attack(&mut self) -> u32 {
+    fn attack(&mut self) -> usize {
         self.attack
     }
 
-    fn step(&mut self) -> u32 {
+    fn step(&mut self) -> usize {
         self.step
     }
 
-    fn value(&self) -> u32 {
+    fn value(&self) -> usize {
         self.value
     }
 
@@ -70,14 +70,14 @@ impl Enemy for EnemySimple {
 
 #[derive(Clone)]
 pub struct EnemyDefense {
-    life: u32,
-    attack: u32,
-    step: u32,
-    value: u32,
+    life: usize,
+    attack: usize,
+    step: usize,
+    value: usize,
 }
 
 impl EnemyDefense {
-    pub fn new(life: u32, attack: u32, step: u32, value: u32) -> Self {
+    pub fn new(life: usize, attack: usize, step: usize, value: usize) -> Self {
         EnemyDefense {
             life,
             attack,
@@ -106,15 +106,15 @@ impl Enemy for EnemyDefense {
         }
     }
 
-    fn attack(&mut self) -> u32 {
+    fn attack(&mut self) -> usize {
         self.attack
     }
 
-    fn step(&mut self) -> u32 {
+    fn step(&mut self) -> usize {
         self.step
     }
 
-    fn value(&self) -> u32 {
+    fn value(&self) -> usize {
         self.value
     }
 
@@ -129,14 +129,14 @@ impl Enemy for EnemyDefense {
 
 #[derive(Clone)]
 pub struct EnemyBig {
-    life: u32,
-    attack: u32,
-    step: u32,
-    value: u32,
+    life: usize,
+    attack: usize,
+    step: usize,
+    value: usize,
 }
 
 impl EnemyBig {
-    pub fn new(life: u32, attack: u32, step: u32, value: u32) -> Self {
+    pub fn new(life: usize, attack: usize, step: usize, value: usize) -> Self {
         EnemyBig {
             life,
             attack,
@@ -164,15 +164,15 @@ impl Enemy for EnemyBig {
         }
     }
 
-    fn attack(&mut self) -> u32 {
+    fn attack(&mut self) -> usize {
         self.attack
     }
 
-    fn step(&mut self) -> u32 {
+    fn step(&mut self) -> usize {
         self.step
     }
 
-    fn value(&self) -> u32 {
+    fn value(&self) -> usize {
         self.value
     }
