+++ music_vs_robot/src/entity/player/consts.rs
@@ -1,24 +1,24 @@
-pub const SAMPLE_HEALTH: u32 = 100;
-pub const SAMPLE_HEALTH_INCREMENT: u32 = 10;
-pub const SAMPLE_COST: u32 = 10;
-pub const SAMPLE_DAMAGE: u32 = 10;
+pub const SAMPLE_HEALTH: usize = 100;
+pub const SAMPLE_HEALTH_INCREMENT: usize = 10;
+pub const SAMPLE_COST: usize = 10;
+pub const SAMPLE_DAMAGE: usize = 10;
 
-pub const THREE_COLUMN_HEALTH: u32 = 100;
-pub const THREE_COLUMN_HEALTH_INCREMENT: u32 = 10;
-pub const THREE_COLUMN_COST: u32 = 10;
-pub const THREE_COLUMN_DAMAGE: u32 = 10;
+pub const THREE_COLUMN_HEALTH: usize = 100;
+pub const THREE_COLUMN_HEALTH_INCREMENT: usize = 10;
+pub const THREE_COLUMN_COST: usize = 10;
+pub const THREE_COLUMN_DAMAGE: usize = 10;
 
-pub const DOUBLE_LIFE_HEALTH: u32 = 100;
-pub const DOUBLE_LIFE_HEALTH_INCREMENT: u32 = 10;
-pub const DOUBLE_LIFE_COST: u32 = 10;
-pub const DOUBLE_LIFE_DAMAGE: u32 = 0;
+pub const DOUBLE_LIFE_HEALTH: usize = 100;
+pub const DOUBLE_LIFE_HEALTH_INCREMENT: usize = 10;
+pub const DOUBLE_LIFE_COST: usize = 10;
+pub const DOUBLE_LIFE_DAMAGE: usize = 0;
 
-pub const THREE_ROW_HEALTH: u32 = 100;
-pub const THREE_ROW_HEALTH_INCREMENT: u32 = 10;
-pub const THREE_ROW_COST: u32 = 10;
-pub const THREE_ROW_DAMAGE: u32 = 0;
+pub const THREE_ROW_HEALTH: usize = 100;
+pub const THREE_ROW_HEALTH_INCREMENT: usize = 10;
+pub const THREE_ROW_COST: usize = 10;
+pub const THREE_ROW_DAMAGE: usize = 0;
 
-pub const SLOW_DOWN_HEALTH: u32 = 100;
-pub const SLOW_DOWN_HEALTH_INCREMENT: u32 = 10;
-pub const SLOW_DOWN_COST: u32 = 10;
-pub const SLOW_DOWN_DAMAGE: u32 = 0;
+pub const SLOW_DOWN_HEALTH: usize = 100;
+pub const SLOW_DOWN_HEALTH_INCREMENT: usize = 10;
+pub const SLOW_DOWN_COST: usize = 10;
+pub const SLOW_DOWN_DAMAGE: usize = 0;
