+++ music_vs_robot/src/entity/player/sample.rs
@@ -3,12 +3,12 @@ use crate::entity::player::{consts::*, player::Player};
 use crate::util::visitor::{Visitable, Visitor};
 
 pub struct PlayerSample {
-    health: u32,
-    level: u32,
+    health: usize,
+    level: usize,
 }
 
 impl PlayerSample {
-    pub fn new(level: u32) -> Self {
+    pub fn new(level: usize) -> Self {
         Self {
             health: SAMPLE_HEALTH + SAMPLE_HEALTH_INCREMENT * level,
             level,
@@ -23,15 +23,15 @@ impl Visitable for PlayerSample {
 }
 
 impl Player for PlayerSample {
-    fn get_health(&self) -> u32 {
+    fn get_health(&self) -> usize {
         self.health
     }
 
-    fn set_health(&mut self, health: u32) {
+    fn set_health(&mut self, health: usize) {
         self.health = health;
     }
 
-    fn get_level(&self) -> u32 {
+    fn get_level(&self) -> usize {
         self.level
     }
 
@@ -40,8 +40,8 @@ impl Player for PlayerSample {
         self.health += SAMPLE_HEALTH_INCREMENT;
     }
 
-    fn get_cost(&self) -> u32 {
-        SAMPLE_COST * (self.level as f32 - 0.75) as u32
+    fn get_cost(&self) -> usize {
+        SAMPLE_COST * (self.level as f32 - 0.75) as usize
     }
 
     fn attack(&self) -> DamagePlayer {
