+++ music_vs_robot/src/playground/cash.rs
@@ -1,16 +1,16 @@
 use crate::util::observer::{Observer, Subject};
 
 pub trait Counter {
-    fn get(&self) -> u32;
+    fn get(&self) -> usize;
     fn clean(&mut self);
 
-    fn add(&mut self, n: u32);
-    fn sub(&mut self, n: u32);
+    fn add(&mut self, n: usize);
+    fn sub(&mut self, n: usize);
 }
 
 pub struct Cash {
     obs: Vec<Box<dyn Observer>>,
-    count: u32,
+    count: usize,
 }
 
 impl Cash {
@@ -23,7 +23,7 @@ impl Cash {
 }
 
 impl Counter for Cash {
-    fn get(&self) -> u32 {
+    fn get(&self) -> usize {
         self.count
     }
 
@@ -31,11 +31,11 @@ impl Counter for Cash {
         self.count = 0;
     }
 
-    fn add(&mut self, n: u32) {
+    fn add(&mut self, n: usize) {
         self.count += n;
     }
 
-    fn sub(&mut self, n: u32) {
+    fn sub(&mut self, n: usize) {
         if n < self.count {
             self.count -= n;
         }
